+++ tests/cli/clitest.cpp
@@ -1,7 +1,7 @@
 // Copyright Glen Knowles 2016 - 2018.
 // Distributed under the Boost Software License, Version 1.0.
 //
-// clitest.cpp - dim test cli
+// clitest.cpp - dimcli test cli
 #include "pch.h"
 #pragma hdrstop