+++ tests/cli/pch.h
@@ -1,7 +1,7 @@
 // Copyright Glen Knowles 2016 - 2017.
 // Distributed under the Boost Software License, Version 1.0.
 //
-// pch.h - dim test cli
+// pch.h - dimcli test cli
 
 #include "dimcli/cli.h"