+++ tools/perf/pch.h
@@ -1,7 +1,7 @@
 // Copyright Glen Knowles 2016 - 2017.
 // Distributed under the Boost Software License, Version 1.0.
 //
-// pch.h - dim test perf
+// pch.h - dimcli test perf
 
 #include "dimcli/cli.h"