+++ tools/perf/perftest.cpp
@@ -1,7 +1,7 @@
 // Copyright Glen Knowles 2016 - 2017.
 // Distributed under the Boost Software License, Version 1.0.
 //
-// perftest.cpp - dim test perf
+// perftest.cpp - dimcli test perf
 #include "pch.h"
 #pragma hdrstop